// src/AdventOfCode.Site/scripts/ts/models/Solution.ts
// Copyright (c) Martin Costello, 2015. All rights reserved.
// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.

export interface Solution {
    year: number;
    day: number;
    solutions: any[];
    visualizations: string[];
    timeToSolve: number;
}